refactor(ui): extract `FormState`, `defaultFormState`, and `buildPatchFromForm` from sheet into `pricingFields.ts` (#6577)
## Summary

Moves `FormState`, `defaultFormState`, `buildPatchFromForm`, and related types (`ScopeRoot`) out of `pricingOverrideSheet.tsx` and into `pricingFields.ts`. This allows the patch-building logic and form state definitions to be unit tested independently without importing the sheet's React component tree.

## Changes

- `FormState`, `ScopeRoot`, `defaultFormState`, and `buildPatchFromForm` now live in `pricingFields.ts` alongside the other shared pricing field metadata.
- `pricingOverrideSheet.tsx` re-exports these from `pricingFields.ts` to preserve the existing public API for consumers.
- Required type imports (`RequestType`, `PricingOverrideMatchType`, `PricingOverridePatch`) were added to `pricingFields.ts` to support the moved logic.

## Type of change

- [ ] Bug fix
- [ ] Feature
- [x] Refactor
- [ ] Documentation
- [ ] Chore/CI

## Affected areas

- [ ] Core (Go)
- [ ] Transports (HTTP)
- [ ] Providers/Integrations
- [ ] Plugins
- [x] UI (React)
- [ ] Docs

## How to test

```sh
cd ui
pnpm i || npm i
pnpm test || npm test
pnpm build || npm run build
```

## Screenshots/Recordings

No visual changes.

## Breaking changes

- [ ] Yes
- [x] No

## Related issues

## Security considerations

None.

## Checklist

- [ ] I read `docs/contributing/README.md` and followed the guidelines
- [ ] I added/updated tests where appropriate
- [ ] I updated documentation where needed
- [ ] I verified builds succeed (Go and UI)
- [ ] I verified the CI pipeline passes locally if applicable
